Update to create new component for dashboard table.
