+++ package.json
@@ -55,6 +55,7 @@
     "webpack-dev-server": "^4.13.2"
   },
   "dependencies": {
+    "@metrostar/comet-data-viz": "1.0.0",
     "@metrostar/comet-extras": "^1.1.0",
     "@metrostar/comet-uswds": "1.2.0",
     "@uswds/uswds": "3.4.1",
